+++ src/commands/manuals/publish.rs
@@ -37,7 +37,11 @@ DESCRIPTION
     within the channel known as the index. For every publish of an ip, the index 
     corresponds to a unique path within the channel that gets created by Orbit.
     A channel's pre-publish and post-publish hooks can get the value for the ip's 
-    index by reading the ORBIT_IP_INDEX environment variable.
+    index by reading the ORBIT_CHANNEL_DIR environment variable.
+    
+    The '--all-public' flag can be used to skip providing the "ip.public" field
+    in the current ip's manifest, granting the assumption that all source files
+    should be public. It cannot be used if the "ip.public" field exists.
     
     By default, this command performs a dry run, which executes all of the steps 
     in the process except for actually posting the ip to its channel(s). 
@@ -52,6 +56,9 @@ OPTIONS
 
     --list
         View available channels and exit
+
+    --all-public
+        Publish with all source files being public
 
 EXAMPLES
     orbit publish