+++ tests/s4/aka1/aka1.sv
@@ -1,0 +1,3 @@
+module aka1;
+
+endmodule;