+++ tests/s4/aka2/aka2.sv
@@ -1,0 +1,3 @@
+module aka2;
+
+endmodule;