+++ CHANGELOG.md
@@ -15,7 +15,13 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 
 # Changelog
 
-## 0.24.1 - unreleased
+## 0.25.0 - unreleased
+
+### Features
+- Adds checks for existence of source files and correct source file visibility before installing and publishing
+- Adds `--all-public` flag to assume source files as public when implicitly private
+- Adds `require-public` entry to configuration files to set default behavior of source files being either private by default or public by default
+- Enforces behavior of source files being private by default
 
 ### Changes
 - Renames `--no-clean` flag to `--keep`, `-k`, to align with conventional linux command-line options